requirements.txt
aiogram[i18n]==3.22.0
httpx
python-dotenv
sqlalchemy
yt-dlp
mutagen
